requirements.txt
langchain
langgraph
langchain-ibm
python-dotenv
duckduckgo-search
langchain-community
beautifulsoup4
lxml
ipython
uv
